pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.86.0",
    "google-generativeai>=0.8.6",
    "imutils>=0.5.4",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "opencv-python-headless>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "scikit-image>=0.26.0",
    "streamlit>=1.55.0",
]
